pyproject.toml
[project]
name = "ingresos"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "gspread>=5.12,<7",
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "openpyxl>=3.1.5",
    "beautifulsoup4>=4.14.3",
    "urllib3>=2.6.3",
    "yfinance>=1.2.0",
    "matplotlib>=3.8.0",
    "numpy>=1.26.0",
]

[tool.uv]
dev-dependencies = []
